#49: update tests
