+++ .vscode/settings.json
@@ -6,6 +6,6 @@
     "**/CVS": true,
     "**/.DS_Store": true,
     "**/Thumbs.db": true,
-    "**/*.mapper.dart": true
+    "**/*.mapper.dart": false
   }
 }
