+++ lib/notes/fields/note_image_field.dart
@@ -1,7 +1,4 @@
-import 'package:dart_mappable/dart_mappable.dart';
-import 'package:noted_models/noted_models.dart';
-
-part 'note_image_field.mapper.dart';
+part of 'note_field.dart';
 
 @MappableClass(discriminatorValue: 'image')
 class NoteImageField extends NoteField<String> with NoteImageFieldMappable {
