+++ lib/notes/fields/note_link_field.dart
@@ -1,7 +1,4 @@
-import 'package:dart_mappable/dart_mappable.dart';
-import 'package:noted_models/noted_models.dart';
-
-part 'note_link_field.mapper.dart';
+part of 'note_field.dart';
 
 @MappableClass(discriminatorValue: 'link')
 class NoteLinkField extends NoteField<String> with NoteLinkFieldMappable {
